[feat/#6-mySelf] useJinary의 코어 로직을 분리, core/jinary.ts 생성
